ADD: Support PTI Three-Winding Transformers
Adds support for Three-winding transformers in PTI files. Transformers
are modeled by creating a "hidden star point bus", and creating a new
transformer branch for each winding of the transformer. For details
of mathematical formulation of the three transformer branches, see
"Power System Stability and Control" by Prabha Kundur, ISBN
0-07-035958-X, Sec. 6.2, Pg. 241, Eqn. 6.72.

Also adds ability to perform unit conversions in the cases that
values of the transformer WIND, R, and X, are not in the base units
defined by the buses and system SBASE. Applies to both two-winding
and three-winding transformers.

Adds unit tests for Three-winding Transformer. Created a Three-winding
transformer system in a PTI file, and used matpower and PowerWorld to
subsequently translate to matpower-type files for comparison.

Updated relevant documentation and CHANGELOG.

Refactors loops inside functions that parse components to eliminate
the enumeration variable (not needed).

Partially adddresses #212 and #231