ADD: Support for transformer magnetizing admittance
Adds support for non-zero magnetizing admittance parameters in PTI files
given via the MAG1 and MAG2 parameters in the transformer entries.
Feature is implemented by a front-side complex admittance, support for
which was added in #230, by setting `g_fr = MAG1` and `b_fr = MAG2`
on its associated transformer branch. For three-winding transformers,
the complex admittance is only added to the first of the three branches
that are created for each of the three windings.

Unit tests are added to check for consistency with PowerWorld PF
solutions, and to ensure OPF feasibility.

Documentation and changelog have been updated.

Added DataStructures to REQUIRE, and import within PowerModels.jl.

Addresses #212 and #231

Closes #255