ADD: Adds key "pti_id" to uniquely identify PTI components
Adds "pti_id" to supported components imported from PTI files. "pti_id"
is an Array of Strings and Ints in each component Dict, the specifics of
which are detailed in the docstrings of the helper functions that import
each type of component.

Added unit tests to check that each supported component has a "pti_id"
of the correct type.

Updated changelog.

Closes #212