UPD: updates network checks for multiphase networks (#302)
* UPD: updates network checks for multiphase networks

Updates the following functions to be compatible with multiphase networks

* check_thermal_limits
* check_branch_directions
* check_transformer_parameters
* check_dcline_limits
* check_voltage_setpoints
* check_cost_functions

Added Base.:* and Base.:/ extensions for MultiPhaseVectors (handles
special broadcast cases).

Fixed issue with Base.broadcast extensions.

Added unit tests for new features.

* UPD: Made phase declaration in warn messages consistent

Added consistency in warning messages in the declaration of the phase
being refered to. If the data is not multiphase, no phase will be
indicated, and if the data is multiphase, the phase will be explicited
named in the warning.

* UPD: change condition for `ph_str`

Changes condition for `ph_str` because it is possible to have a multiphase
network with only a single phase.