FIX: PSSE string splitting (#471)
* FIX: regex update

* FIX: Improper ' stripping if space before, after

Single quotes were not being properly stripped in cases where there was space before or after the '.

Fix typo in commen split regex

* FIX: Single quote inside PSSE string must be escaped

Previously, PowerModels supported the existance of a single single-quote inside of string fields that were indicated by surrounding single-quotes. However, this cannot be supported, and in fact, other PSSE parsers (see PowerWorld) also do not support this. Instead we will require this quotes to be escaped.

* ADD: Error if uneven number of quotes detected

Adds an error message for if an uneven number of quotes is detected, and associated unit test.

* ADD: unit test for string stripping bugfix

* DOC: Update changelog